requirement.txt
flask
python-dotenv
langchain
langchain-core
langgraph
langchain-google-genai
google-generativeai
requests
